requirements.txt
pytorch
torchvision
h5py
tqdm